Move class LoginFlowType to `api` package
